refactor(deps): move _optional_deps from scqubits/ui/ to scqubits/utils/
The previous home in scqubits/ui/ forced scqubits/core/circuit.py and scqubits/core/circuit_internals/sym_methods.py to import from scqubits.ui, inverting the natural ui -> core dependency direction. scqubits/utils/ is already a cross-cutting utility sink that both core/ and ui/ import from, so moving the optional-deps module there resolves the inversion without introducing new edges.

No behavior change: _optional_deps.py itself is unchanged; the 16 existing import sites and the test fixture pick up the new path.
