+++ scqubits/utils/_optional_deps.py
@@ -5,7 +5,7 @@ flags plus the optional module references (``v``, ``display``, etc.)
 that the Explorer subsystem reaches for.  Files that previously
 duplicated the ``try: import ipyvuetify as v; except ImportError:
 _HAS_IPYVUETIFY = False; else: _HAS_IPYVUETIFY = True`` boilerplate
-can now ``from scqubits.ui._optional_deps import v, _HAS_IPYVUETIFY``.
+can now ``from scqubits.utils._optional_deps import v, _HAS_IPYVUETIFY``.
 
 When a dependency is missing, the corresponding name is set to
 ``None`` and the ``_HAS_*`` flag is ``False``.  Call sites that
